Enhance CSS styles across multiple files, add new calendar event functionality, improve user feedback on ticket creation, and implement confirmation dialogs for reporting and account deletion actions. Update JSON error messages and sitemap to include new routes.

+++ Sources/BackOffice/Processus/delete_category.php
@@ -9,10 +9,24 @@ if($is_admin != 0){
         $id = intval($_GET['id']);
 
         try {
+
+            $count_number_of_categories = $pdo->prepare("SELECT COUNT(id) FROM CATEGORY");
+            $count_number_of_categories->execute();
+            $count = $count_number_of_categories->fetchColumn();
+
+            if($count <= 1){
+                //header("Location: ../database_gestion.php?error=last_category");
+                echo "<script>window.location.href = '../database_gestion.php?error=last_category';</script>";
+                exit();
+            }
+
             $stmt = $pdo->prepare("DELETE FROM CATEGORY WHERE id = :id");
             $stmt->bindParam(':id', $id, PDO::PARAM_INT);
-
             $stmt->execute();
+
+            $delete_all_petitions_inside_this_category = $pdo->prepare("DELETE FROM PETITION WHERE category = :id");
+            $delete_all_petitions_inside_this_category->bindParam(':id', $id, PDO::PARAM_INT);
+            $delete_all_petitions_inside_this_category->execute();
 
             header("Location: ../database_gestion.php");
             exit();
